+++ tests/testthat/test-mask.R
@@ -36,3 +36,31 @@ test_that("basic", {
   plot(textmask[1])
   plot(listmask[1])
 })
+
+## custom mask from vectors ----
+test_that("custom mask from vectors", {
+  stimuli <- demo_stim("test")
+  
+  stimuli[1] %>%
+    crop_tem() %>%
+    resize(2) %>%
+  draw_tem(pt.shape = "index", pt.size = 10) %>% plot()
+  
+  list_mask <- list(list(
+    c(71:75, 50, 56, 78:82),
+    c(82,49),
+    c(49:47),
+    c(47:46),
+    c(46:44),
+    c(44, 71)
+  ))
+  text_mask <- lapply(list_mask, lapply, paste, collapse = ",") %>%
+    lapply(paste, collapse = ";") %>%
+    paste(collapse = ":")
+  
+  t <- mask(stimuli[1], text_mask)
+  l <- mask(stimuli[1], list_mask)
+  
+  expect_equal(c(f_multi = 0), 
+               image_comp(t, l, scale = TRUE))
+})